modified with your suggestions
